+++ game_patch/graphics/d3d11/gr_d3d11_context.cpp
@@ -529,7 +529,7 @@ namespace df::gr::d3d11
             current_color_.blue / 255.0f,
             current_color_.alpha / 255.0f,
         };
-        data.alpha_test = current_alpha_test_ ? (1.0f / 255.0f) : 0.0f;
+        data.alpha_test = current_alpha_test_ ? current_alpha_test_threshold_ : 0.0f;
         if (!current_fog_allowed_ || !gr::screen.fog_mode) {
             data.fog_far = std::numeric_limits<float>::infinity();
             data.fog_color = {0.0f, 0.0f, 0.0f};